Removed various commands that are no longer needed or to be used. These where STFC specific, and most would be accessing database tables that either no longer exist or have been refactored.
Most of the removed commands dealt with capturing and/or displaying score data for users.

Renamed the help command filename from datahelp to just data. This reflects the new generic name given to this command already in a previous commit.

The other two files removed by this commit are lcars and officercheetsheet, which are also STFC specific commands, the former relating to the lcars project which had already become defunct, whilsts the latter was a helper for STFC in-game officer characters.